refactor: move SelectedFilters in SortAndFilter
